scrutinizer tips
