+++ dev/tests/app/model/Win/DesignPattern/DemoSingleton.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Win\DesignPattern;
+
+class DemoSingleton extends SuperDemoSingleton {
+
+	public $value = 20;
+
+}
