+++ dev/tests/app/model/Win/DesignPattern/SuperDemoSingleton.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace Win\DesignPattern;
+
+class SuperDemoSingleton {
+
+	use Singleton;
+
+	public $value = 10;
+
+	public function getMyClass() {
+		return $this->getClassDi();
+	}
+
+}
